Create 19I-1750 Fateh Muhammad
I have written and published a blog "Introduction to Docker | part1" on Medium. This blog is meant for Beginners, so anyone with just computing background should be able to understand the importance of Docker and how Docker provides an isolated environment for different applications in the form of Docker Containers to run without any dependency issues. I have also covered, how the architecture of Docker is similar to Client Server Architecture and how the Docker engine manages Containers.